Renamed SceneGraph to SceneHierarchy
